requirements.txt
fastapi
pydantic
sqlalchemy
passlib
bcrypt==3.2.0
python-jose[cryptography]